requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
discord==2.3.2
discord.py==2.4.0
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
yarl==1.12.1
